+++ usr.sbin/makefs/hammer2/hammer2_inode.c
@@ -1716,7 +1716,7 @@ hammer2_inode_chain_ins(hammer2_inode_t *ip)
 			error = 0;
 		if (error) {
 			kprintf("hammer2: backend unable to "
-				"insert inode %p %ld\n", ip, ip->meta.inum);
+				"insert inode %p %ld\n", ip, (long)ip->meta.inum);
 			/* XXX return error somehow? */
 		}
 	}
@@ -1755,7 +1755,7 @@ hammer2_inode_chain_des(hammer2_inode_t *ip)
 			error = 0;
 		if (error) {
 			kprintf("hammer2: backend unable to "
-				"delete inode %p %ld\n", ip, ip->meta.inum);
+				"delete inode %p %ld\n", ip, (long)ip->meta.inum);
 			/* XXX return error somehow? */
 		}
 	}
@@ -1836,8 +1836,8 @@ vflush(struct mount *mp, int rootrefs, int flags)
 		count_after++;
 	hammer2_spin_unex(&pmp->inum_spin);
 
-	printf("%s: total inode %ld -> %ld\n",
-	    __func__, count_before, count_after);
+	printf("%s: total inode %jd -> %jd\n",
+	    __func__, (intmax_t)count_before, (intmax_t)count_after);
 
 	assert(count_before >= count_after);
 	count_delta = count_before - count_after;
@@ -1846,7 +1846,8 @@ vflush(struct mount *mp, int rootrefs, int flags)
 		if (hammer2_debug & 0x80000000)
 			assert(0);
 		else
-			printf("%s: %ld inode freed\n", __func__, count_delta);
+			printf("%s: %jd inode freed\n", __func__,
+			    (intmax_t)count_delta);
 	}
 
 	return 0;
